backend/requirements.txt
flask
flask-cors
chromadb
openai
python-dotenv
pandas
pinecone
